requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.8
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
greenlet==1.0.0
idna==3.1
idna-ssl==1.1.0
importlib-metadata==3.10.0
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
psycopg2-binary
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.24
typing-extensions==3.7.4.3
SQLAlchemy-Utils==0.36.8
yarl==1.6.3
zipp==3.4.1
pytest==6.2.3
pytest-aiohttp==0.3.0